+++ tests/test_config.py
@@ -1,25 +1,10 @@
 """ unit tests for config.py """
 
-import unittest
-
 from flask import current_app
-from app import create_app, db
+from .test_base import BaseTestCase
 
-class ConfigTestCase(unittest.TestCase):
+class ConfigTestCase(BaseTestCase):
     """ unit tests for basic config and setup"""
-    def setUp(self):
-        """initial setup before a test is run """
-        self.app = create_app('testing')
-        self.app_context = self.app.app_context()
-        self.app_context.push()
-        db.create_all()
-
-    def tearDown(self):
-        """executes  after a test is run """
-        db.session.remove()
-        db.drop_all()
-        self.app_context.pop()
-
     def test_app_exists(self):
         """ test if app exists"""
         self.assertFalse(current_app is None)
